put all the widgets container in widgets folder
